Add menu order feature on Business Option

+++ database/migrations/2017_11_22_052705_create_business_options_table.php
@@ -23,7 +23,7 @@ class CreateBusinessOptionsTable extends Migration {
             $table->text('content', 65535)->nullable();
             $table->string('element', 191)->nullable();
             $table->text('tooltip', 65535)->nullable();
-            $table->tinyInteger('menu_order')->unique()->nullable();
+            $table->integer('menu_order')->unique()->nullable();
 			$table->integer('weight')->nullable();
             $table->boolean('show_everywhere')->default(true);
 			$table->timestamps();
